TEST(Products): ProductEventTest 가 정상 동작하지않아서 수정을하였습니다. 이벤트가 발행이 안되는 이슈를 가지고 있습니다.
